+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `find` verb - filter items matching criteria, returns `FindResult(indices, reason)`
+
 ## [0.2.0] - 2026-03-16
 
 ### Added